work thus far before going to pc to work on finsihing main.py for the big vision of nexus theory to become a galatic empire
