+++ requirements.txt
@@ -7,7 +7,7 @@ numpy>=1.23.0
 pandas>=1.5.0
 scipy>=1.9.0
 matplotlib>=3.6.0
-seaborn>=0.12.0
+# seaborn>=0.12.0
 
 # Machine Learning & LLM Development
 torch>=2.0.0
